Remove coordinate reprocessing code
This was added in an attempt to deal with OCR noise:

https://github.com/LibraryOfCongress/chronam/commit/87d0216fdc88ede6a35fed222d551b1f97d94a8f

Unfortunately it also strips non-English text: languages
such as Arabic would be completely removed and non-ASCII
characters on the beginning or end of a line would also be
silently stripped.

This commit avoids data loss but does it would likely also
be necessary to review the original impetus for this request
and see whether we need to make the JavaScript highlighting
code smarter about matching words ignoring punctuation.