Newspaper feed: load in less than the timeout period
Removing the last duplicate calculation of the most recent issue value
shaves one query per title off of the load time (it is now constant at 3
queries) and this is further reduced by only loading the fields which
are displayed in the feed. Combined this keeps the generation time under
the load-balancer timeout.

Before: 2m45.268s

After:  0m8.634s