Refactor newspaper statistics display
This caused performance and stability issues because it meant that every Django template render could potentially trigger both some expensive database queries and a Request to Solr.

Only one of the defined variables was used on more than one page so all of the database queries used on tabs.html are moved into the view which renders that template.

The total page count is display on most pages on the site but since it’s supposed to be globally consistent it has been moved into an AJAX request for the relevant JSON data and only displays when that returns successfully (previous data as stale as a year old would be displayed).