Fix custom SQL reporting query
This returned an error on newer versions of MySQL which follow the SQL standard requirement that the columns in ORDER BY be in the GROUP BY as well — previously this worked by coincidence because the differing column was the formatted date so the values matched.

For performance, this does the GROUP BY on the native stored value rather than the calculated string representation.