Remove RDF content negotiation decorator
This was lightly used since our generated links use the trailing-slash
URL for HTML and the .rdf suffixed URL for RDF and the people who did
hit on it either reported it as a bug or changed their URLs because it
didn't implement the Vary header following the HTTP standard and thus
whether you got HTML or RDF depended in part on whether previous
requests had primed the cache.

Since all of the links and documentation use the non-negotiated .rdf we
can simplify the logic by removing the negotiated views and letting
Django’s normal APPEND_SLASH behavior handle any legacy usage.